Changes...Please Read Description
The only things I changed in the util classes was spelling and stuff, so
if the thing freaks out then it's on me.

Finished the storyline, just working on the flow of the game. We'll
probably never get to test the EndGame class since we'll never get
there.

Alex, please work on how leveling up affects the damage and health of a
player.